Moved date to first column, subject to last
This means that we do not have to truncate the subject, because the
buffer is set to truncate lines and will do it for us.

Therefore, we do not have to calculate the subject width.

Furthermore, we do not have to nm-refresh on if the window width
changes but the window height does not.

We still refresh if the window height changes.  Obviously we must if
the window height grows.  Even if it shrinks we need to keep the
results list in sync with the height.