Bug fix for auto wakeup
The run-at-time function takes an etime (i.e., result of encode-time)
as input, that is a list of two ints.  Internally we use lists that
can have more elements but for timing purposes we only look at the
first two elements.  run-at-time does not like this, so when we pass
it a time we pass only the first two elements.