Workaround for display-buffer-below-selected new behavior
The behavior of display-buffer-below-selected changed in commit
6a5d9e9a3247566bf69575444e2a1f75f5b3e021
in the emacs-24 branch of git://git.savannah.gnu.org/emacs.git

The old behavior (as in the parent commit
73283e1e5f880749da43b9b0b572272b2358de16) is that if the buffer already
has a window, it is simply reused.

The new behavior is that if the buffer already has a window, it is not
reused, and the selected window is split and the buffer is shown in the
new window below.

I think this is a bug in emacs, but I've put in a workaround in any
case.