Correct message search sort order
Notmuch does not search by message, it sorts first by thread.
I've added a binary that invokes the notmuch library to provide the
correct sorting.
Install the binary with

    go install github.com/tjim/nevermore/go/nevermore

The emacs mail client will use the binary if installed in exec-path, or
you can set nm-nevermore-command to the absolute path.